Remove header blockchain and weight verification, provides no value here
