+++ tests/full_node/test_block_store.py
@@ -108,6 +108,7 @@ class TestBlockStore:
             db_filename.unlink()
             db_filename_2.unlink()
             db_filename_3.unlink()
+            b.shut_down()
             raise
 
         await connection.close()
@@ -116,6 +117,7 @@ class TestBlockStore:
         db_filename.unlink()
         db_filename_2.unlink()
         db_filename_3.unlink()
+        b.shut_down()
 
     @pytest.mark.asyncio
     async def test_deadlock(self):
