+++ tests/full_node/test_blockchain.py
@@ -62,6 +62,7 @@ class TestGenesisBlock:
         assert bc1.get_next_min_iters(bc1.genesis) > 0
 
         await connection.close()
+        bc1.shut_down()
 
 
 class TestBlockValidation:
@@ -522,6 +523,7 @@ class TestBlockValidation:
         assert (b.get_next_min_iters(blocks[27])) == (b.get_next_min_iters(blocks[26]))
 
         await connection.close()
+        b.shut_down()
 
 
 class TestReorgs:
@@ -556,6 +558,7 @@ class TestReorgs:
         assert b.get_current_tips()[0].height == 119
 
         await connection.close()
+        b.shut_down()
 
     @pytest.mark.asyncio
     async def test_reorg_from_genesis(self):
@@ -600,6 +603,7 @@ class TestReorgs:
         assert result == ReceiveBlockResult.ADDED_TO_HEAD
 
         await connection.close()
+        b.shut_down()
 
     @pytest.mark.asyncio
     async def test_lca(self):
@@ -629,6 +633,7 @@ class TestReorgs:
         assert b.lca_block.header_hash == blocks[0].header_hash
 
         await connection.close()
+        b.shut_down()
 
     @pytest.mark.asyncio
     async def test_find_fork_point(self):
@@ -672,6 +677,7 @@ class TestReorgs:
         )
         assert b.lca_block.data == blocks[4].header.data
         await connection.close()
+        b.shut_down()
 
     @pytest.mark.asyncio
     async def test_get_header_hashes(self):
@@ -691,3 +697,4 @@ class TestReorgs:
         assert header_hashes == [block.header_hash for block in blocks]
 
         await connection.close()
+        b.shut_down()
